requirements.txt
# Pip requirements file

requests
paho-mqtt
six
futures
